+++ webdev/test/integration_test.dart
@@ -30,7 +30,7 @@ void main() {
 
   var invalidRanges = <String, List<String>>{
     'build_runner': ['0.7.13+1', '0.9.0'],
-    'build_web_compilers': ['0.3.5', '0.4.0']
+    'build_web_compilers': ['0.3.5', '0.5.0']
   };
 
   for (var command in ['build', 'serve']) {
@@ -119,7 +119,7 @@ name: sample
               } else {
                 assert(entry.key == 'build_web_compilers');
                 webCompilersVersion = version;
-                supportedRange = '^0.3.6';
+                supportedRange = '>=0.3.6 <0.5.0';
               }
 
               await d.file('pubspec.yaml', '''
@@ -242,7 +242,7 @@ dependencies:
 }
 
 const _supportedBuildRunnerVersion = '0.8.2';
-const _supportedWebCompilersVersion = '0.3.6';
+const _supportedWebCompilersVersion = '0.4.0';
 
 String _pubspecLock(
     {String runnerVersion: _supportedBuildRunnerVersion,