+++ webdev/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.1
 
+- Support `package:build_web_compilers` `v0.4.0`.
 - Always pass the argument `--fail-on-severe` to the `build` command.
 - Exit with an error if unsupported arguments are passed to `build` command.